Rename loggers class file
