+++ trainerdex_discord_bot/loggers.py
